test: move agency JSON->XML round-trip out of the xml-only job
Reading SDMX-JSON imports jsonschema (doc_validation.py) at module load,
regardless of validate=, so an xml-marked test that reads a JSON fixture
fails under the json-less xml-only CI job. Move it to test_general_writer.py
(not in conftest PATH_RULES), which the single-extra jobs skip and the
all-extras build matrix runs. Mirrors the categorisation round-trip test.
